requirements.txt
streamlit==1.24.0
requests==2.31.0
pandas==2.0.3
numpy==1.25.1
python-dotenv==1.0.0 